requirements.txt
flask
flask_cors
requests
openai
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-flask
azure-monitor-opentelemetry-exporter
gunicorn
Flask-HTTPAuth